moved shot class to game_sprites.py
